Mark 3 GtkOverlay functions are post-3.18.
These functions were added in 3.18, so don't try to use them
with earlier versions:
gtk_overlay_reorder_overlay()
gtk_overlay_get_overlay_pass_through()
gtk_overlay_set_overlay_pass_through()
