+++ gtk/gtk.go
@@ -5806,24 +5806,6 @@ func (v *Overlay) AddOverlay(widget IWidget) {
 	C.gtk_overlay_add_overlay(v.native(), widget.toWidget())
 }
 
-// ReorderOverlay() is a wrapper around gtk_overlay_reorder_overlay().
-func (v *Overlay) ReorderOverlay(child IWidget, position int) {
-	C.gtk_overlay_reorder_overlay(v.native(), child.toWidget(), C.gint(position))
-}
-
-// GetOverlayPassThrough() is a wrapper around
-// gtk_overlay_get_overlay_pass_through().
-func (v *Overlay) GetOverlayPassThrough(widget IWidget) bool {
-	c := C.gtk_overlay_get_overlay_pass_through(v.native(), widget.toWidget())
-	return gobool(c)
-}
-
-// SetOverlayPassThrough() is a wrapper around
-// gtk_overlay_set_overlay_pass_through().
-func (v *Overlay) SetOverlayPassThrough(widget IWidget, passThrough bool) {
-	C.gtk_overlay_set_overlay_pass_through(v.native(), widget.toWidget(), gbool(passThrough))
-}
-
 /*
  * GtkPaned
  */
